+++ R/draws.R
@@ -102,7 +102,7 @@ draws_any.rvec <- function(x, na_rm = FALSE) {
 #' Credible Intervals from Random Draws
 #'
 #' Summarise the distribution of random draws
-#' in an `rvec`, using a simple credible interval.
+#' in an `rvec`, using  credible intervals.
 #'
 #' @section Warning:
 #'
@@ -114,12 +114,14 @@ draws_any.rvec <- function(x, na_rm = FALSE) {
 #' `my_df$ci <- draws_ci(my_rvec)`
 #'
 #' However, creating columns in
-#' this way can corrupt data frames.
+#' this way can corrupt an ordinary data frames.
 #' For safer options,
 #' see the examples below.
 #'
 #' @inheritParams draws_all
-#' @param width A number, where `0 < width <= 1`.
+#' @param width Width(s) of credible interval(s).
+#' One or more numbers greater than 0
+#' and less than or equal to 1.
 #' Default is `0.975`.
 #' @param prefix String to be added to the
 #' names of columns in the result.
@@ -158,6 +160,8 @@ draws_any.rvec <- function(x, na_rm = FALSE) {
 #' x <- rvec(m)
 #' x
 #' draws_ci(x)
+#' draws_ci(x, width = c(0.5, 0.99))
+#' draws_ci(x, prefix = "results")
 #'
 #' ## results from 'draws_ci'
 #' ## assigned to a data frame
@@ -185,31 +189,39 @@ draws_ci.rvec <- function(x,
                           width = 0.95,
                           prefix = NULL,
                           na_rm = FALSE) {
-    x_str <- deparse1(substitute(x))
-    check_width(width)
-    has_prefix <- !is.null(prefix)
-    if (has_prefix)
-        check_str(prefix, x_arg = "prefix")
-    check_flag(na_rm)
-    probs <- make_probs(width)
-    m <- field(x, "data")
-    if (nrow(m) == 0L)
-        ans <- stats::quantile(double(), probs = probs)
-    else {
-        ans <- matrixStats::rowQuantiles(m,
-                                         probs = probs,
-                                         na.rm = na_rm,
-                                         drop = FALSE)
-        ans <- matrix_to_list_of_cols(ans)
-    }
-    nms <- c(".lower", ".mid", ".upper")
-    if (has_prefix)
-        nms <- paste0(prefix, nms)
-    else
-        nms <- paste0(x_str, nms)
-    names(ans) <- nms
-    ans <- tibble::tibble(!!!ans)
-    ans
+  x_str <- deparse1(substitute(x))
+  check_width(width)
+  has_prefix <- !is.null(prefix)
+  if (has_prefix)
+    check_str(prefix, x_arg = "prefix")
+  check_flag(na_rm)
+  probs <- make_probs(width)
+  m <- field(x, "data")
+  if (nrow(m) == 0L)
+    ans <- stats::quantile(double(), probs = probs)
+  else {
+    ans <- matrixStats::rowQuantiles(m,
+                                     probs = probs,
+                                     na.rm = na_rm,
+                                     drop = FALSE)
+    ans <- matrix_to_list_of_cols(ans)
+  }
+  n <- length(width)
+  lower <- rep(".lower", times = n)
+  upper <- rep(".upper", times = n)
+  if (n > 1L) {
+    s <- seq_len(n - 1L)
+    lower[-1L] <- paste0(lower[-1L], s)
+    upper[-n] <- paste0(upper[-n], rev(s))
+  }
+  nms <- c(lower, ".mid", upper)
+  if (has_prefix)
+    nms <- paste0(prefix, nms)
+  else
+    nms <- paste0(x_str, nms)
+  names(ans) <- nms
+  ans <- tibble::tibble(!!!ans)
+  ans
 }
 
 ## HAS_TESTS